refactor: Move DbPassPhrase to release source set and update passphrase generation
This commit moves the `DbPassPhrase` class from the main source set to the release source set.
It also updates the passphrase generation:
- Uses `SecureRandom()` instead of `SecureRandom.getInstance("SHA1PRNG")` for API levels below O.
- Removes the loop that ensures the passphrase does not contain 0, as this is not strictly necessary for security and can be resource-intensive.
- Adds error handling for `GeneralSecurityException` and `IOException` when getting or generating the passphrase, logging the error and returning null.
