--- core/data/src/main/kotlin/eg/edu/cu/csds/icare/core/data/local/db/DbPassPhrase.kt
@@ -1,53 +1,0 @@
-package eg.edu.cu.csds.icare.core.data.local.db
-
-import android.content.Context
-import androidx.security.crypto.EncryptedFile
-import androidx.security.crypto.MasterKey
-import org.koin.core.annotation.Single
-import java.io.File
-import java.security.SecureRandom
-
-@Single
-class DbPassPhrase(
-    private val context: Context,
-) {
-    fun getPassphrase(): ByteArray {
-        val file = File(context.filesDir, "passphrase.bin")
-        val encryptedFile =
-            EncryptedFile
-                .Builder(
-                    context,
-                    file,
-                    MasterKey
-                        .Builder(context)
-                        .setKeyScheme(MasterKey.KeyScheme.AES256_GCM)
-                        .build(),
-                    EncryptedFile.FileEncryptionScheme.AES256_GCM_HKDF_4KB,
-                ).build()
-
-        return if (file.exists()) {
-            encryptedFile.openFileInput().use { it.readBytes() }
-        } else {
-            generatePassphrase().also { passphrase ->
-                encryptedFile.openFileOutput().use { it.write(passphrase) }
-            }
-        }
-    }
-
-    private fun generatePassphrase(): ByteArray {
-        val random =
-            if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
-                SecureRandom.getInstanceStrong()
-            } else {
-                SecureRandom.getInstance("SHA1PRNG")
-            }
-        val result = ByteArray(size = 32)
-
-        random.nextBytes(result)
-        while (result.contains(0)) {
-            random.nextBytes(result)
-        }
-
-        return result
-    }
-}
